Extracted job handlers
