+++ shop/jobs/Job.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace shop\jobs;
+
+abstract class Job implements \zhuravljov\yii\queue\Job
+{
+    public function execute($queue): void
+    {
+        $listener = $this->resolveHandler();
+        $listener($this, $queue);
+    }
+
+    private function resolveHandler(): callable
+    {
+        return [\Yii::createObject(static::class . 'Handler'), 'handle'];
+    }
+}
